moved class to hot-button.jsx
