+++ react-hot-button-component/src/index.jsx
@@ -1,27 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import hotButton from 'hot-button';
-
-let counter = 1;
-
-class CustomButton extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {isClicked: false};
-        this.handleClick = this.handleClick.bind(this);
-    }
-
-    
-
-    handleClick() {
-        this.setState({isClicked: counter++})
-    }
-
-    render() {
-        <hotButton />
-    }
-}
+import CustomButton from 'hot-button';
 
 ReactDOM.render(
     <CustomButton />,
